+++ src/lib/components/cosmos-evm-txs/data.ts
@@ -6,9 +6,10 @@ import {
   useEvmTxsDataJsonRpc,
   useTxsByAddressPaginationSequencer,
 } from "lib/services/tx";
+import { isUndefined } from "lodash";
 import { useEffect, useState } from "react";
 
-export const useContractDetailsEvmTxs = (address: BechAddr20) => {
+export const useCosmosEvmTxs = (address: BechAddr20) => {
   const [isInitialLoading, setIsInitialLoading] = useState(true);
 
   const [evmTxs, setEvmTxs] = useState<TxDataWithTimeStampJsonRpc[]>();
@@ -38,13 +39,19 @@ export const useContractDetailsEvmTxs = (address: BechAddr20) => {
     data: newEvmTxsData,
     isError: isNewEvmTxsDataError,
     isFetching: isNewEvmTxsDataFetching,
+    remove: removeNewEvmTxsData,
   } = useEvmTxsDataJsonRpc(
     newEvmTxHashes?.filter((tx) => tx !== null) as string[],
     !isCosmosTxsFetching && !isNewEvmHashesFetching
   );
 
   useEffect(() => {
-    if (!cosmosTxs || !newEvmTxHashes || !newEvmTxsData) return;
+    if (
+      isUndefined(cosmosTxs) ||
+      isUndefined(newEvmTxHashes) ||
+      isUndefined(newEvmTxsData)
+    )
+      return;
 
     const newEvmTxs: TxDataWithTimeStampJsonRpc[] = [];
     newEvmTxHashes.forEach((evmTxHash, index) => {
@@ -55,11 +62,14 @@ export const useContractDetailsEvmTxs = (address: BechAddr20) => {
         });
     });
 
+    setEvmTxs((evmTxs ?? []).concat(newEvmTxs));
+    setCosmosTxsCount(cosmosTxsCount + cosmosTxs.items.length);
+    setPaginationKey(cosmosTxs.pagination.nextKey);
     setIsInitialLoading(false);
 
-    setEvmTxs((evmTxs ?? []).concat(newEvmTxs));
-    setPaginationKey(cosmosTxs.pagination.nextKey);
-    setCosmosTxsCount(cosmosTxsCount + cosmosTxs.items.length);
+    // Note: cosmosTxs and newEvmTxHashes are undefined when paginationKey is changed
+    // So, we need to remove newEvmTxsData every time we fetch new data
+    removeNewEvmTxsData();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [newEvmTxsData]);