+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#1339](https://github.com/alleslabs/celatone-frontend/pull/1339) Filter slingy tx on lite version
 - [#1333](https://github.com/alleslabs/celatone-frontend/pull/1333) Support NFT collection details, NFT details on WASM and EVM
 
 ### Improvements